Refactoring into chef_local.rb

+++ runtime/chef_local.rb
@@ -3,13 +3,17 @@ require 'rubygems'
 require 'json'
 require 'tempfile'
 
+abort "Syntax : chef_local.rb host [additionnal chef cookbooks & roles path]" if ARGV == 0
+
 server = ARGV[0]
-localrolepath = ARGV[1]
+additionnal_path = ARGV[1]
 user = "chef"
 
-abort "Incorrect Syntax, you should use it like this : chef_local.rb servername localrolepath \n Locale role path is the folder containings your custom role and configuration override" unless ARGV.length == 2
-
-puts "Running chef with local cookbooks : on #{user}@#{server} and local role path : #{localrolepath} "
+if additionnal_path
+  puts "Running chef with local cookbooks : on #{user}@#{server} with additionnal_path #{additionnal_path}"
+else
+  puts "Running chef with local cookbooks : on #{user}@#{server} without additionnal_path"
+end
 
 def exec_local cmd
   begin
@@ -19,8 +23,8 @@ def exec_local cmd
   end
 end
  
-["../master-chef", localrolepath].each do |dir|
-  exec_local "rsync --delete -avh --exclude=.git #{dir}/ #{user}@#{server}:/tmp/#{File.basename(dir)}/"
+["../master-chef", additionnal_path].each do |dir|
+  exec_local "rsync --delete -avh --exclude=.git #{dir}/ #{user}@#{server}:/tmp/#{File.basename(dir)}/" if dir
 end
 
 exec_local "scp #{user}@#{server}:/etc/chef/local.json /tmp/"
@@ -31,7 +35,7 @@ f = Tempfile.new 'local.json'
 f.write JSON.pretty_generate(json)
 f.close
 
-exec_local "scp #{f.path}  #{user}@#{server}:/tmp/local.json"
+exec_local "scp #{f.path} #{user}@#{server}:/tmp/local.json"
 exec_local "ssh #{user}@#{server} MASTER_CHEF_CONFIG=/tmp/local.json /etc/chef/rbenv_sudo_chef.sh -c /etc/chef/solo.rb"
